requirements.txt
appdirs==1.4.2
clearbit==0.1.5
click==6.7
colorama==0.3.7
Django==1.10.6
django-cors-headers==2.0.2
django-rq==0.9.4
djangorestframework==3.5.4
djangorestframework-jwt==1.9.0
docopt==0.6.2
ez-setup==0.9
feedparser==5.2.1
gitsome==0.6.0
numpydoc==0.6.0
packaging==16.8
pkg-resources==0.0.0
ply==3.10
post==0.0.0
prompt-toolkit==1.0.13
Pygments==2.2.0
PyJWT==1.4.2
pyparsing==2.1.10
pytz==2016.10
redis==2.10.5
requests==2.13.0
rq==0.7.1
setupfiles==0.0.0
six==1.10.0
type==0.0.1
uritemplate==3.0.0
uritemplate.py==3.0.2
validate-email==1.3
wcwidth==0.1.7
